requirements.txt
alabaster==0.7.13 ; python_version >= "3.10" and python_version < "4.0"
arrow==1.2.3 ; python_version >= "3.10" and python_version < "4.0"
babel==2.11.0 ; python_version >= "3.10" and python_version < "4.0"
beautifulsoup4==4.11.2 ; python_version >= "3.10" and python_version < "4.0"
binaryornot==0.4.4 ; python_version >= "3.10" and python_version < "4.0"
bs4==0.0.1 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
chardet==5.1.0 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
cookiecutter==2.1.1 ; python_version >= "3.10" and python_version < "4.0"
docutils==0.19 ; python_version >= "3.10" and python_version < "4.0"
furo==2022.12.7 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
imagesize==1.4.1 ; python_version >= "3.10" and python_version < "4.0"
jinja2-time==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
lxml==4.9.2 ; python_version >= "3.10" and python_version < "4.0"
markdown-it-py==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
mdit-py-plugins==0.3.4 ; python_version >= "3.10" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.10" and python_version < "4.0"
myst-parser==0.18.1 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.10" and python_version < "4.0"
pygments==2.14.0 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
python-magic==0.4.27 ; python_version >= "3.10" and python_version < "4.0"
python-slugify==8.0.0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
rich==13.3.1 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
snowballstemmer==2.2.0 ; python_version >= "3.10" and python_version < "4.0"
soupsieve==2.4 ; python_version >= "3.10" and python_version < "4.0"
sphinx-basic-ng==1.0.0b1 ; python_version >= "3.10" and python_version < "4.0"
sphinx==5.3.0 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-applehelp==1.0.4 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-devhelp==1.0.2 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-htmlhelp==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-jsmath==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-qthelp==1.0.3 ; python_version >= "3.10" and python_version < "4.0"
sphinxcontrib-serializinghtml==1.1.5 ; python_version >= "3.10" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.10" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.10" and python_version < "4"
zundler @ git+https://github.com/AdrianVollmer/Zundler@HEAD ; python_version >= "3.10" and python_version < "4.0"
